jobs/requirements.txt
asgiref==3.8.1
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.1
psycopg2-binary== 2.9.9
Faker==28.0.0
django-filter==24.3
Markdown==3.7